requirements.txt
absl-py==1.4.0
accelerate
aiofiles==22.1.0
aiosqlite==0.18.0
alembic
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.6.2
blis==0.7.10
boto3==1.28.38
botocore==1.31.38
Bottleneck==1.3.5
bpemb==0.3.4
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
chart-studio==1.1.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.2
colorlover==0.3.0
comm==0.1.2
confection==0.1.1
conllu==4.5.3
construct==2.5.3
contourpy==1.0.5
cryptography==41.0.3
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
daal4py==2023.1.1
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
dlib==19.24.2
dnspython==2.4.2
email-validator==2.0.0.post2
en-core-web-sm @https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
entrypoints==0.4
essentia==2.1b6.dev1034
et-xmlfile==1.1.0
executing==0.8.3
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.3
Faker==19.3.1
fastjsonschema==2.16.2
fer==22.5.1
ffmpeg==1.4
filelock==3.12.2
flair
Flask==2.3.3
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.25.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
fuzzywuzzy==0.18.0
gast==0.4.0
gdown==4.4.0
gensim==4.3.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
gunicorn==21.2.0
h5py==3.9.0
haystack==0.42
huggingface-hub
hyperopt==0.2.7
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.8
install==1.3.5
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
Janome==0.5.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.6
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-server==1.23.4
jupyter-ydoc==0.2.4
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.22.0
kaleido==0.2.1
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
lazy_loader==0.3
Levenshtein==0.21.1
libclang==16.0.6
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.1
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==10.1.0
moviepy==1.0.3
mpld3==0.3
mpmath==1.3.0
munkres==1.1.4
murmurhash==1.0.9
mysqlclient==2.0.3
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numba==0.57.1
numexpr==2.8.4
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai-whisper
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandastable==0.13.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pdfkit==1.0.0
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
plotly==5.9.0
ply==3.11
pptree==3.1
preshed==3.0.8
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic_core
pydub==0.25.1
pygal==3.0.0
Pygments==2.15.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-ptrace==0.9.8
pytorch_revgrad
pytz==2022.7
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.4.2
QtPy==2.2.0
rapidfuzz==3.2.0
redis==5.0.0
regex==2022.7.9
requests==2.29.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.2
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
segtok==1.5.11
Send2Trash==1.8.0
sentence-transformers
sentencepiece==0.1.99
sip==6.6.2
six==1.16.0
smart-open==6.3.0
sniffio==1.2.0
soundfile==0.12.1
soupsieve==2.4
spacy
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
spotipy==2.23.0
SQLAlchemy
sqlitedict==2.1.0
srsly==2.4.7
stack-data==0.2.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
terminado==0.17.1
textblob==0.17.1
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.3.3
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch
torchaudio
torchvision
tornado==6.2
tqdm==4.65.0
traitlets== 5.7.1
transformer-smaller-training-vocab
transformers==4.33.2
triton==2.0.0
typer==0.9.0
typing_extensions
urllib3==1.26.16
URLObject==2.4.3
wasabi==1.1.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.3.7
widgetsnbextension==4.0.5
Wikipedia-API==0.6.0
wordcloud==1.9.2
wrapt==1.15.0
WTForms==3.0.1
xgboost==1.7.6
xlrd==2.0.1
y-py==0.5.9
ypy-websocket==0.8.2
